Remove 'image.tag' values in favor of Chart.yaml appVersion
This updates the appVersion values for all of the Crucible apps to the latest published versions. We have been behind in keeping the 'image.tag' values up to date, and the idea with this change is that we can apply CI to this process down the road as we release new versions of Crucible apps.

This change does not impact users who have custom 'image.tag' values in their own deployments. It just makes the default values for each app more current.